sd-sagemaker-deployment/requirements.txt
diffusers[torch]==0.24.0
transformers
accelerate
safetensors
torch
Pillow
controlnet-aux